analytics/requirements.txt
-i https://pypi.org/simple
attrs==20.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
boto3==1.14.56
botocore==1.17.56
cached-property==1.5.1
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
click==6.7
configparser==3.8.1; python_version >= '2.6'
datadotworld[pandas]==1.7.0
datapackage==1.15.0
docutils==0.15.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
et-xmlfile==1.0.1
flask==1.1.2
google-api-core[grpc]==1.22.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth==1.21.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-cloud-secret-manager==1.0.0
googleapis-common-protos[grpc]==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ijson==3.1.1
importlib-metadata==1.7.0; python_version < '3.8'
isodate==0.6.0
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jdcal==1.4.1
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonlines==1.2.0
jsonpointer==2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonschema==3.2.0
linear-tsv==1.1.0
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
numpy==1.16.4
openpyxl==3.0.5; python_version >= '3.6'
pandas==0.24.2
protobuf==3.13.0
psycopg2==2.8.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyrsistent==0.16.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
requests==2.24.0
rfc3986==1.4.0
rsa==4.6; python_version >= '3.5'
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.3.19; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tableschema==1.19.3
tabulator==1.52.3
unicodecsv==0.14.1
urllib3==1.25.10; python_version != '3.4'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
xlrd==1.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
zipp==3.1.0; python_version >= '3.6'
